Conflicts:
	java/com.sap.sse.security.ui/src/main/java/com/sap/sse/security/ui/shared/UserDTO.java

+++ java/com.sap.sse.security.ui/src/main/java/com/sap/sse/security/ui/server/UserManagementServiceImpl.java
@@ -42,6 +42,7 @@ import com.sap.sse.security.shared.Ownership;
 import com.sap.sse.security.shared.OwnershipAnnotation;
 import com.sap.sse.security.shared.QualifiedObjectIdentifier;
 import com.sap.sse.security.shared.Role;
+import com.sap.sse.security.shared.RoleDefinition;
 import com.sap.sse.security.shared.RoleImpl;
 import com.sap.sse.security.shared.SecurityUser;
 import com.sap.sse.security.shared.UnauthorizedException;
@@ -118,11 +119,22 @@ public class UserManagementServiceImpl extends RemoteServiceServlet implements U
     }
 
     @Override
-    public void updateRoleDefinition(RoleDefinitionDTO roleDefinitionWithNewProperties) {
-        // FIXME an additional check is needed to verify that a user may only grant/revoke permissions that he owns
-        // In case of role permissions this means the user needs to own an unqualified version of the permission because
-        // any user may own the role with any qualification
-        SecurityUtils.getSubject().checkPermission(SecuredSecurityTypes.ROLE_DEFINITION.getStringPermissionForObjects(DefaultActions.UPDATE, roleDefinitionWithNewProperties.getId().toString()));
+    public void updateRoleDefinition(RoleDefinitionDTO roleDefinitionWithNewProperties) throws UnauthorizedException {
+        SecurityUtils.getSubject().checkPermission(SecuredSecurityTypes.ROLE_DEFINITION.getStringPermissionForObjects(
+                DefaultActions.UPDATE, roleDefinitionWithNewProperties.getId().toString()));
+        
+        RoleDefinition existingRole = getSecurityService().getRoleDefinition(roleDefinitionWithNewProperties.getId());
+        if (existingRole == null) {
+            throw new UnauthorizedException("Role does not exist");
+        }
+        Set<WildcardPermission> addedPermissions = new HashSet<>(roleDefinitionWithNewProperties.getPermissions());
+        addedPermissions.removeAll(existingRole.getPermissions());
+        
+        if (!getSecurityService().hasUserAllWildcardPermissionsForAlreadyRealizedQualifications(existingRole, addedPermissions)) {
+            throw new UnauthorizedException("Not permitted to grant permissions for role "
+                    + roleDefinitionWithNewProperties.getName());
+        }
+        
         getSecurityService().updateRoleDefinition(roleDefinitionWithNewProperties);
     }
 
@@ -242,8 +254,8 @@ public class UserManagementServiceImpl extends RemoteServiceServlet implements U
     }
 
     @Override
-    public UserGroup createUserGroup(String name, String nameOfTenantOwner) throws UnauthorizedException, UserGroupManagementException {
-        return getSecurityService().setOwnershipCheckPermissionForObjectCreationAndRevertOnError(nameOfTenantOwner,
+    public UserGroup createUserGroup(String name) throws UnauthorizedException, UserGroupManagementException {
+        return getSecurityService().setOwnershipCheckPermissionForObjectCreationAndRevertOnError(
                 SecuredSecurityTypes.USER_GROUP, name, name, () -> {
                     UUID newTenantId = UUID.randomUUID();
                     UserGroup userGroup;
@@ -255,7 +267,7 @@ public class UserManagementServiceImpl extends RemoteServiceServlet implements U
                     getSecurityService().setOwnership(
                             SecuredSecurityTypes.USER_GROUP.getQualifiedObjectIdentifier(newTenantId.toString()),
                             getSecurityService().getCurrentUser(),
-                            getSecurityService().getUserGroupByName(nameOfTenantOwner), name);
+                            getSecurityService().getDefaultTenantForCurrentUser(), name);
                     final Map<SecurityUser, SecurityUser> fromOriginalToStrippedDownUser = new HashMap<>();
                     final Map<UserGroup, UserGroup> fromOriginalToStrippedDownUserGroup = new HashMap<>();
                     return securityDTOFactory.createUserGroupDTOFromUserGroup(userGroup, fromOriginalToStrippedDownUser,
@@ -355,13 +367,13 @@ public class UserManagementServiceImpl extends RemoteServiceServlet implements U
     }
 
     public UserDTO createSimpleUser(String username, String email, String password, String fullName, String company,
-            String localeName, String validationBaseURL, String tenantOwnerName)
+            String localeName, String validationBaseURL)
             throws UserManagementException, MailException, UnauthorizedException {
-        final UserImpl u = getSecurityService().setOwnershipCheckCreatePermissionAndRevertOnError(tenantOwnerName,
+        final UserImpl u = getSecurityService().setOwnershipCheckPermissionForObjectCreationAndRevertOnError(
                 SecuredSecurityTypes.USER, username, username, () -> {
                     try {
                         return getSecurityService().createSimpleUser(username, email, password, fullName, company,
-                                getLocaleFromLocaleName(localeName), validationBaseURL, tenantOwnerName);
+                                getLocaleFromLocaleName(localeName), validationBaseURL);
                     } catch (UserManagementException | UserGroupManagementException e) {
                         logger.log(Level.SEVERE, "Error creating user " + username, e);
                         throw new UserManagementException(e.getMessage());
@@ -435,12 +447,15 @@ public class UserManagementServiceImpl extends RemoteServiceServlet implements U
     public SuccessInfo setRolesForUser(String username,
             Iterable<Triple<UUID, String, String>> roleDefinitionIdAndTenantQualifierNameAndUsernames)
             throws UnauthorizedException {
-        // FIXME an additional check is needed to verify that a user may only grant/revoke permissions that he owns
-        // The current user needs to have all permissions of a role for the specific tenant qualification or an unqualified version
-        if (SecurityUtils.getSubject().isPermitted(
-                SecuredSecurityTypes.USER.getStringPermissionForObjects(UserActions.GRANT_PERMISSION, username))
-                && SecurityUtils.getSubject().isPermitted(SecuredSecurityTypes.USER
-                        .getStringPermissionForObjects(UserActions.REVOKE_PERMISSION, username))) {
+        final boolean isUserPermittedToGrantPermissionsForOtherUser = SecurityUtils.getSubject().isPermitted(
+                SecuredSecurityTypes.USER.getStringPermissionForObjects(UserActions.GRANT_PERMISSION, username));
+        final boolean isUserPermittedToRevokePermissionsForOtherUser = SecurityUtils.getSubject().isPermitted(SecuredSecurityTypes.USER
+                .getStringPermissionForObjects(UserActions.REVOKE_PERMISSION, username));
+        if (!isUserPermittedToGrantPermissionsForOtherUser
+                && !isUserPermittedToRevokePermissionsForOtherUser) {
+            return new SuccessInfo(false, "Not permitted to grant or revoke permissions for user " + username,
+                    /* redirectURL */null, null);
+        } else {
             User u = getSecurityService().getUserByName(username);
             if (u == null) {
                 return new SuccessInfo(false, "User does not exist.", /* redirectURL */ null, null);
@@ -467,12 +482,29 @@ public class UserManagementServiceImpl extends RemoteServiceServlet implements U
             Set<Role> roleDefinitionsToRemove = new HashSet<>();
             Util.addAll(u.getRoles(), roleDefinitionsToRemove);
             Util.removeAll(rolesToSet, roleDefinitionsToRemove);
-            for (Role roleToRemove : roleDefinitionsToRemove) {
-                getSecurityService().removeRoleFromUser(u, roleToRemove);
+            if (!roleDefinitionsToRemove.isEmpty() && !isUserPermittedToRevokePermissionsForOtherUser) {
+                return new SuccessInfo(false, "Not permitted to revoke permissions for user " + username,
+                        /* redirectURL */null, null);
             }
             Set<Role> rolesToAdd = new HashSet<>();
             Util.addAll(rolesToSet, rolesToAdd);
             Util.removeAll(u.getRoles(), rolesToAdd);
+            if (!rolesToAdd.isEmpty() && !isUserPermittedToGrantPermissionsForOtherUser) {
+                return new SuccessInfo(false, "Not permitted to grant permissions for user " + username,
+                        /* redirectURL */null, null);
+            }
+            for (Role roleToAdd : rolesToAdd) {
+                for (WildcardPermission permissionOfRoleToAdd : roleToAdd.getPermissions()) {
+                    if (!getSecurityService().hasCurrentUserMetaPermission(permissionOfRoleToAdd, roleToAdd.getQualificationAsOwnership())) {
+                        return new SuccessInfo(false,
+                                "Not permitted to grant role " + roleToAdd.getName() + " for user " + username,
+                                /* redirectURL */null, null);
+                    }
+                }
+            }
+            for (Role roleToRemove : roleDefinitionsToRemove) {
+                getSecurityService().removeRoleFromUser(u, roleToRemove);
+            }
             for (Role roleToAdd : rolesToAdd) {
                 getSecurityService().addRoleForUser(u, roleToAdd);
             }
@@ -481,8 +513,6 @@ public class UserManagementServiceImpl extends RemoteServiceServlet implements U
             final UserDTO userDTO = securityDTOFactory.createUserDTOFromUser(u, getSecurityService());
             return new SuccessInfo(true, message, /* redirectURL */null,
                     new Pair<UserDTO, UserDTO>(userDTO, getAllUser()));
-        } else {
-            throw new UnauthorizedException("Not permitted to grant permissions to user");
         }
     }
     
@@ -503,10 +533,15 @@ public class UserManagementServiceImpl extends RemoteServiceServlet implements U
 
     @Override
     public SuccessInfo setPermissionsForUser(String username, Iterable<WildcardPermission> permissions) throws UnauthorizedException {
-        if (SecurityUtils.getSubject().isPermitted(
-                SecuredSecurityTypes.USER.getStringPermissionForObjects(UserActions.GRANT_PERMISSION, username))
-                && SecurityUtils.getSubject().isPermitted(SecuredSecurityTypes.USER
-                        .getStringPermissionForObjects(UserActions.REVOKE_PERMISSION, username))) {
+        final boolean isUserPermittedToGrantPermissionsForOtherUser = SecurityUtils.getSubject().isPermitted(
+                SecuredSecurityTypes.USER.getStringPermissionForObjects(UserActions.GRANT_PERMISSION, username));
+        final boolean isUserPermittedToRevokePermissionsForOtherUser = SecurityUtils.getSubject().isPermitted(SecuredSecurityTypes.USER
+                .getStringPermissionForObjects(UserActions.REVOKE_PERMISSION, username));
+        if (!isUserPermittedToGrantPermissionsForOtherUser
+                && !isUserPermittedToRevokePermissionsForOtherUser) {
+            return new SuccessInfo(false, "Not permitted to grant or revoke permissions for user " + username,
+                    /* redirectURL */null, null);
+        } else {
             User u = getSecurityService().getUserByName(username);
             if (u == null) {
                 return new SuccessInfo(false, "User does not exist.", /* redirectURL */null, null);
@@ -514,26 +549,34 @@ public class UserManagementServiceImpl extends RemoteServiceServlet implements U
             Set<WildcardPermission> permissionsToRemove = new HashSet<>();
             Util.addAll(u.getPermissions(), permissionsToRemove);
             Util.removeAll(permissions, permissionsToRemove);
-            for (WildcardPermission permissionToRemove : permissionsToRemove) {
-                getSecurityService().removePermissionFromUser(username, permissionToRemove);
+            if (!permissionsToRemove.isEmpty() && !isUserPermittedToRevokePermissionsForOtherUser) {
+                return new SuccessInfo(false, "Not permitted to revoke permissions for user " + username,
+                        /* redirectURL */null, null);
             }
             Set<WildcardPermission> permissionsToAdd = new HashSet<>();
             Util.addAll(permissions, permissionsToAdd);
             Util.removeAll(u.getPermissions(), permissionsToAdd);
+            if (!permissionsToAdd.isEmpty() && !isUserPermittedToGrantPermissionsForOtherUser) {
+                return new SuccessInfo(false, "Not permitted to grant permissions for user " + username,
+                        /* redirectURL */null, null);
+            }
             for (WildcardPermission permissionToAdd : permissionsToAdd) {
-                OwnershipAnnotation currentUsersOwnerShip = getSecurityService()
-                        .getOwnership(getSecurityService().getCurrentUser().getIdentifier());
-                if (getSecurityService().hasCurrentUserMetaPermission(permissionToAdd,
-                        currentUsersOwnerShip.getAnnotation())) {
-                    getSecurityService().addPermissionForUser(username, permissionToAdd);
+                if (!getSecurityService().hasCurrentUserMetaPermission(permissionToAdd, null)) {
+                    return new SuccessInfo(false,
+                            "Not permitted to grant permission " + permissionToAdd + " for user " + username,
+                            /* redirectURL */null, null);
                 }
+            }
+            for (WildcardPermission permissionToRemove : permissionsToRemove) {
+                getSecurityService().removePermissionFromUser(username, permissionToRemove);
+            }
+            for (WildcardPermission permissionToAdd : permissionsToAdd) {
+                getSecurityService().addPermissionForUser(username, permissionToAdd);
             }
             final String message = "Set roles " + permissions + " for user " + username;
             final UserDTO userDTO = securityDTOFactory.createUserDTOFromUser(u, getSecurityService());
             return new SuccessInfo(true, message, /* redirectURL */null,
                     new Pair<UserDTO, UserDTO>(userDTO, getAllUser()));
-        } else {
-            throw new UnauthorizedException("Not permitted to grant or revoke permissions for user "+username);
         }
     }
